+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## NEXT VERSION
 
+- perf: optimize `unflatten` to not use recursion
+
 # v1.2.1 (2019-05-01)
 
 - fix: scrollbar size don't updated in SSR